refactor(home): update follow refactor code base
